backend/face-swapping-service/pyproject.toml
[project]
name = "face-swapping-service"
version = "0.1.0"
description = "face swapping service"
requires-python = ">=3.9,<3.12"
dependencies = [
    "aibox-image-lib",
    "pyzmq>=25.0.0",
    "msgpack>=1.0.0",
    "zmq>=0.0.0",
    "torch>=2.8.0",
    "pillow>=11.3.0",
    "opencv-python>=4.11.0.86",
    "numpy>=1.24.0",
    "insightface>=0.7.3",
    "onnxruntime>=1.16.0",
    "huggingface_hub>=0.16.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0"]

[tool.uv.sources]
aibox-image-lib = { path = "../aibox-image-lib", editable = true }

[project.scripts]
service = "face_swapping_service.main:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
